requirements.txt
Django==3.1.3
django-debug-toolbar==2.2

black
isort==4.3.21
pycodestyle==2.5.0